promotion related code review
1. Split and rename "branched_promotion*" to "classchg-*" to get aligned to efx proc names.

2. Decompile classchg-postconfirm code to complete the total bmpromote code.
