+++ include/bmbattle.h
@@ -99,7 +99,7 @@ enum {
     BATTLE_CONFIG_ARENA = (1 << 5),
     BATTLE_CONFIG_REFRESH = (1 << 6),
     BATTLE_CONFIG_MAPANIMS = (1 << 7),
-    BATTLE_CONFIG_PROMOTION_CHAR100 = (1 << 8),
+    BATTLE_CONFIG_PROMOTION_PREP = (1 << 8),
     BATTLE_CONFIG_DANCERING = (1 << 9),
 };
