+++ src/bmguide.c
@@ -1,0 +1,13 @@
+#include "global.h"
+
+const char gProcName_E_guideSub[] = "E_guideSub";
+const char gProcName_E_guMenu1ReWrite[] = "E_guMenu1ReWrite";
+const char gProcName_E_guMenu2ReWriteFirst[] = "E_guMenu2ReWriteFirst";
+const char gProcName_E_guMenu2ReWriteUp[] = "E_guMenu2ReWriteUp";
+const char gProcName_E_guMenu2ReWriteDown[] = "E_guMenu2ReWriteDown";
+const char gProcName_E_guMess3ReWrite[] = "E_guMess3ReWrite";
+const char gProcName_E_Guide[] = "E_Guide";
+
+// Pad the data here so that the linker script doesn't need to specify the
+// exact location of data after this file.
+const u8 TextFiller[] = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 };
