+++ src/prep_menuproc.c
@@ -16,7 +16,7 @@
 #include "bmudisp.h"
 #include "bm.h"
 #include "prepscreen.h"
-#include "branched_promotion.h"
+#include "classchg.h"
 
 s8 CheckInLinkArena();
 
@@ -54,7 +54,7 @@ void DoPromoteAnimForChar100(struct Proc08A184B4 *proc)
     SetWinEnable(0, 0, 0);
     
     sub_802F598(unit, -1, 0);
-    gBattleStats.config = BATTLE_CONFIG_PROMOTION | BATTLE_CONFIG_PROMOTION_CHAR100;
+    gBattleStats.config = BATTLE_CONFIG_PROMOTION | BATTLE_CONFIG_PROMOTION_PREP;
     gBattleActor.weaponBefore = 0;
     gBattleTarget.weaponBefore = 0;
     BeginBattleAnimations();
